+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[1.3.0] - 2018-06-07
+---------------------
+###### Added
+- Appearance validation.
+
+###### Changed
+- Updated Enketo libraries.
+
 [1.2.2] - 2018-05-01
 ---------------------
 ##### Changed